Improve / add page header
Wraps the introductory paragraph in a div.page-header, and adds a h1
before it. To make all this still fit on the screen, I had to reduce the
default rows of the stacktrace. This shouldn’t be a problem as desktop
browsers (i. e., where you have the screen real estate) allow you to
resize textareas.